backend/requirements.txt
flask
flask-mysqldb
flask-jwt-extended
pycryptodome
python-dotenv